move create flag env to env service
