+++ packages/backend/src/flags/flags.service.ts
@@ -1,8 +1,6 @@
 import { Injectable } from '@nestjs/common';
-import camelcase from 'camelcase';
 import { StrategyService } from '../strategy/strategy.service';
 import { PrismaService } from '../prisma.service';
-import { FlagAlreadyExists } from './errors';
 import { FlagStatus } from './flags.status';
 import { FlagHitsRetrieveDTO, PopulatedFlagEnv } from './types';
 import { FieldRecord } from '../strategy/types';
@@ -76,40 +74,6 @@ export class FlagsService {
     `;
 
     return hits;
-  }
-
-  async createFlag(envId: string, name: string, description: string) {
-    const flagKey = camelcase(name);
-
-    const existingFlag = await this.prisma.flagEnvironment.findFirst({
-      where: {
-        environmentId: envId,
-        flag: {
-          key: flagKey,
-        },
-      },
-    });
-
-    if (existingFlag) {
-      throw new FlagAlreadyExists();
-    }
-
-    const flag = await this.prisma.flag.create({
-      data: {
-        name,
-        description,
-        key: flagKey,
-      },
-    });
-
-    await this.prisma.flagEnvironment.create({
-      data: {
-        flagId: flag.uuid,
-        environmentId: envId,
-      },
-    });
-
-    return flag;
   }
 
   async deleteFlag(envId: string, flagId: string) {
